requirements.txt
aiohttp
aiosignal
attrs
beautifulsoup4
certifi
charset-normalizer
Brotli
multidict
python-dotenv
rich
soupsieve
yarl
gspread
google-auth
pandas
dotenv
aiohttp
aiosignal
attrs
beautifulsoup4
certifi
charset-normalizer
frozenlist
idna
multidict
python-dotenv
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib